secure-api-backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.24.0
redis>=5.0.1
pydantic[email]>=2.6.3
pydantic-settings==2.1.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.9
bcrypt>=4.0.1
python-dotenv>=1.0.0
httpx==0.24.1
cryptography>=41.0.5
typing-extensions==4.9.0
asyncio==3.4.3
gunicorn>=21.0.0
loguru>=0.7.0
pycryptodome>=3.19.0
openai>=1.14.0 